requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
TextBlob==0.15.3
SpeechRecognition==3.8.1
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
python-dotenv==1.0.1
Werkzeug==3.0.1
reportlab==3.6.8
PyAudio==0.2.11
mysql-connector-python==8.0.27
pydub==0.25.1
wave==0.0.2
